test: move SmartlingWireMock into a separate `test-utils` module
