+++ pom.xml
@@ -149,6 +149,7 @@
 
     <!-- list in alphabetical order, except api-sdk-all -->
     <modules>
+        <module>samrtling-api-test-utils</module>
         <module>smartling-api-commons</module>
         <module>smartling-issues-api</module>
         <module>smartling-jobs-api</module>
