+++ samrtling-api-test-utils/pom.xml
@@ -2,24 +2,16 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
+    <artifactId>smartling-api-test-utils</artifactId>
+    <packaging>jar</packaging>
+    <version>0.14.1-SNAPSHOT</version>
+    <name>Smartling API Test Utils</name>
+
     <parent>
         <groupId>com.smartling.api</groupId>
         <artifactId>smartling-sdk-parent</artifactId>
         <version>0.14.1-SNAPSHOT</version>
     </parent>
-
-    <artifactId>smartling-files-api</artifactId>
-    <version>0.14.1-SNAPSHOT</version>
-    <name>Smartling Content Files API</name>
-
-    <build>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <filtering>true</filtering>
-            </resource>
-        </resources>
-    </build>
 
     <dependencies>
         <dependency>
@@ -27,16 +19,11 @@
             <artifactId>smartling-api-commons</artifactId>
             <version>0.14.1-SNAPSHOT</version>
         </dependency>
+
         <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-            <version>3.8.1</version>
-        </dependency>
-        <dependency>
-            <groupId>com.smartling.api</groupId>
-            <artifactId>smartling-api-test-utils</artifactId>
-            <version>0.14.1-SNAPSHOT</version>
-            <scope>test</scope>
+            <groupId>com.github.tomakehurst</groupId>
+            <artifactId>wiremock</artifactId>
+            <version>2.26.0</version>
         </dependency>
     </dependencies>
 </project>
