+++ smartling-api-commons/pom.xml
@@ -11,4 +11,13 @@
         <artifactId>smartling-sdk-parent</artifactId>
         <version>0.14.1-SNAPSHOT</version>
     </parent>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.smartling.api</groupId>
+            <artifactId>smartling-api-test-utils</artifactId>
+            <version>0.14.1-SNAPSHOT</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 </project>
